requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
python-telegram-bot==13.12
environs==9.5.0
lxml==4.9.1
